Extract import-time value coercion helpers to importer_value_coercion
Sprint 4, PR #5.  Third peel of modules/importer.py.  Moves the
~179-line cluster of tiny, single-purpose helpers that turn
heterogeneous user YAML values into the tight, DB-friendly shapes
prepare_import_payload stores.

None of these functions know or care about the wider import
pipeline; they're all pure transformations on individual values.
Ideal candidates for their own module so the coercion rules can
be reviewed and unit-tested in isolation.

Public-ish API (all leading '_' -- internal but re-exported from
modules/importer for test monkeypatching and for prepare_import_payload
which stays behind):

Type coercion:
- _coerce_import_bool       -- YAML-wide truthy/falsy strings
                               (regression-guarded by
                               tests/test_importer_edge_cases)
- _coerce_import_bool_text  -- Returns 'true'/'false'/'' for DB round-trip
- _coerce_import_int        -- int + numeric string -> int | None
- _coerce_import_string_list -- scalar/list -> list[str]

Template-var / overlay-meta introspection:
- _collect_template_keys
- _collect_dynamic_child_field_specs
- _collect_overlay_source_override_keys
- _has_template_string_list_values

Serialization:
- _serialize_playlist_import_value
- _serialize_dynamic_child_mapping_value

This is a PURE MOVE.  No coercion or serialization logic changed.
Every returned value is byte-identical.

Compatibility:
- All 10 names re-exported from modules/importer so:
  * The 12+ call sites inside prepare_import_payload resolve
    unchanged.
  * tests/test_importer_edge_cases which uses
    importer._coerce_import_bool keeps working with zero changes.

Layout note:
- In the source file _coerce_import_bool sat 384 lines below the
  other 9 coercion helpers (at L704 vs L154-320).  This was
  a historical artifact from the two-def-bug killed in PR #1504.
  Both are now co-located in the new module.

- _build_attribute_sets (~80 lines at L423-502) is NOT part of
  this extraction -- it's about parsing attribute config
  metadata, not coercing user values.  Stays with its sole
  caller prepare_import_payload.

Dependency direction: importer -> importer_value_coercion
(one way).  Zero imports from importer, so no circular imports
possible.

Impact:
- modules/importer.py: 1777 -> 1598 (-179 / -10.1%)
- modules/importer_value_coercion.py: NEW ~250 lines

All 1285 tests pass.  Ruff + black clean.
